requirements.txt
absl-py==0.7.1
aiohttp==3.5.4
aiohttp-cors==0.7.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cycler==0.10.0
Cython==0.29.13
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
html5lib==0.999999999
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
imageio==2.6.1
imutils==0.5.3
incremental==16.10.1
jedi==0.14.1
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
multidict==4.5.2
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.0
object-detection-0.1
opencv-python==4.1.0.25
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pbr==3.1.1
pickleshare==0.7.5
Pillow==6.1.0
playsound==1.2.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
pycparser==2.19
Pygments==2.4.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.2
pyRFC3339==1.1
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
pywinpty==0.5.5
pyzmq==18.1.0
reportlab==3.4.0
requests==2.18.4
#requests-unixsocket==0.1.5
requests-unixsocket
scikit-learn==0.21.2
scipy==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
slim==0.1.0
flask==1.1.2
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.11.2
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
yarl==1.3.0
